reservation app logic

+++ alembic/versions/b45b3ceb2d9b_created_user_table.py
@@ -30,7 +30,7 @@ def upgrade():
     sa.Column('username', sa.String(), nullable=False),
     sa.Column('is_admin', sa.Boolean(), nullable=True),
     sa.Column('is_active', sa.Boolean(), nullable=True),
-    sa.Column('amount_paid', sa.Float(), nullable=True),
+    sa.Column('amount_paid', sa.Integer(), nullable=True),
     sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
     sa.PrimaryKeyConstraint('id')
     )
